Test case splitted into two files
